- Добавлена возможность работы при отсутствии соединения с БД - Проведена декомпозиция кода - Класс с методами выведен в отдельный модуль (adpy-11)

+++ adpy-11/diplom/main5.py
@@ -1,172 +1,12 @@
 import vk
+from user_class import User
 import private_settings
 from datetime import datetime
 import time
 import operator
 import json
 import sql
-
-
-class User():
-    """Клас описывает основные параметры пользователя, которые будем сохранять в базу"""
-
-    def __init__(self, name, api):
-        self.start_age = None
-        self.finish_age = None
-        self.top3 = None
-        self.api = api
-        self.name = name
-        self.common = ''
-        self.get_data()
-        self.groups_get()
-
-    def get_data(self):
-        self.user_info = self.api.users.get(user_ids=self.name, fields='bdate, sex, city, \
-        books, music, interests, photo_max')
-        self.id = self.user_info[0]['id']
-        self.sex = self.user_info[0]['sex']
-        self.photo_max = self.user_info[0]['photo_max']
-        if 'city' not in self.user_info[0]:
-            self.city = False
-        else:
-            self.city = self.user_info[0]['city']
-        if 'music' not in self.user_info[0]:
-            self.music = False
-        else:
-            self.music = self.user_info[0]['music']
-        if 'books' not in self.user_info[0]:
-            self.books = False
-        else:
-            self.books = self.user_info[0]['books']
-        if 'interests' not in self.user_info[0]:
-            self.interests = False
-        else:
-            self.interests = self.user_info[0]['interests']
-        if 'bdate' not in self.user_info[0]:
-            self.bdate = False
-            self.age = False
-        else:
-            self.bdate = self.user_info[0]['bdate']
-            try:
-                birthday = datetime.strptime(self.bdate, "%d.%m.%Y")
-                age = birthday.year
-                self.age = datetime.now().year - age
-            except:
-                self.age = False
-
-    def groups_get(self):
-        time.sleep(0.33)
-        print('Получить группы пользователя ', self.id)
-        try:
-            self.groups = self.api.groups.get(user_id=self.id)['items']
-            print('Получен список групп пользователя', self.id, ' ', len(self.groups))
-        except Exception as err:
-            self.groups = []
-        return self.groups
-
-    def get_mutual(self, target):
-        time.sleep(0.33)
-        try:
-            self.mutual_friends = self.api.friends.getMutual(source_uid=self.id, target_uid=target)
-        except Exception as err:
-            print('Профиль ', self.id, ' закрыт, данные не получены')
-            self.mutual_friends = False
-
-    def calc_kpi(self, target):
-        """Метод находит общее между пользователем и объектом исследования на основе следующей системы весов
-        общий друг              10 баллов
-        совпадение возраста     7 баллов
-        общая группа            5 баллов
-        совпадение по музыке    3 балла
-        совпадение по книге     2 балла
-        совпадение по интересам 1 балл
-        """
-        kpi = 0
-        try:
-            for friend in self.mutual_friends:
-                kpi += 10
-                self.common += ('начислено 10 баллов за общего другана - ' + str(self.mutual_friends) + '\n')
-        except:
-            pass  # нет друзей
-        if self.age == target.age:
-            kpi += 7
-            self.common += ('начислено 7 баллов за совпадение по возрасту' + '\n')
-        # поиск общих групп
-        for group in target.groups:
-            if group in self.groups:
-                kpi += 5
-                self.common += (' начислено 5 баллов за общую группу - ' + str(group) + '\n')
-        # поиск совпадений по музыке
-        if self.music and target.music:
-            pattern = self.music.split(',')
-            for i in pattern:
-                if len(i) > 2:
-                    if i in target.music:
-                        kpi += 3
-                        self.common += ('начислено 3 балла за совпадение по музыке - ' + i + '\n')
-
-        # поиск совпадений по книгам
-        if self.books and target.books:
-            pattern = self.books.split(',')
-            for i in pattern:
-                if len(i) > 2:
-                    if i in target.books:
-                        kpi += 2
-                        self.common += (' начислено 2 балла за совпадение по книгам - ' + i + '\n')
-
-        # поиск совпадений по интересам
-        if self.interests and target.interests:
-            pattern = self.interests.split(',')
-            for i in pattern:
-                if len(i) > 2:
-                    if i in target.interests:
-                        kpi += 1
-                        self.common += ('начислен 1 балл за совпадение по интересам - ' + i + '\n')
-        self.kpi = kpi
-        self.common += ('-' * 20 + '\nВсего совпадений на ' + str(kpi) + ' баллов.')
-
-    def info(self, api):
-        """ Метод возвращает массив данных о юзере"""
-        user = api.users.get(user_id=self.id)
-        return user[0]
-
-    def search_users(self):
-        """ Метод получает список пользователей подходящих под базовый запрос (город, диапазон возраста, статус, пол) """
-        if self.sex == 1:
-            sex_ = 2
-        else:
-            sex_ = 1
-        if self.city:
-            candidates = self.api.users.search(city=self.city['id'], sex=sex_, age_from=self.start_age,
-                                               age_to=self.finish_age, status=[1], count=50)
-        else:
-            candidates = self.api.users.search(sex=sex_, age_from=self.start_age,
-                                               age_to=self.finish_age, status=[1], count=50)
-        candidates_items = candidates['items']
-        candidates_ = list()
-        for i in candidates_items:
-            candidates_.append(i['id'])
-        return candidates_
-
-    def show_result(self):
-        tmp_list = []
-        self.url = ('https://vk.com/id' + str(self.id))
-        try:
-            for i in self.top3:
-                s = i['sizes']
-                for j in s:
-                    if j['type'] == 'x':
-                        tmp_list.append(j['url'])
-
-        except:
-            tmp_list.append('Профиль закрыт, есть только фото профиля')
-            tmp_list.append(self.photo_max)
-        answer = dict()
-        answer['id'] = str(self.id)
-        answer['url'] = str(self.url)
-        answer['photos'] = tmp_list
-        self.answer = answer
-        return answer
+import file_base
 
 
 def get_token():
@@ -202,7 +42,7 @@ def get_target(api):
 
 
 def get_age(user):
-    """Функция проверяет, указан ли возраст, и просит ввести вручную, если данных нет"""
+    """ Функция проверяет, указан ли возраст, и просит ввести вручную, если данных нет """
     try:
         birthday = datetime.strptime(user.bdate, "%d.%m.%Y")
         age = birthday.year
@@ -213,6 +53,29 @@ def get_age(user):
     user.age = age
     user.start_age = age - 2
     user.finish_age = age + 2
+
+
+def get_city(user):
+    """ Функция проверяет, указан ли город, и просит ввести вручную, если данных нет """
+    print('Функция определения города вызвана')
+    print(f'{user.city=}')
+    while True:
+        if user.city == False:
+            city = input('Введите город проживания пользователя: ')
+            list_of_cityes = user.api.database.getCities(country_id=1, q=city)
+            if list_of_cityes['count'] == 0:
+                print('Населенного пункта с таким наименованием несуществует. Попробуйте снова.')
+                continue
+            break
+    print(list_of_cityes)
+    last_city = list_of_cityes['count']
+    if list_of_cityes['count'] > 21:
+        last_city = 21
+    for i in range(1, last_city):
+        print(i, ' - ', list_of_cityes['items'][i])
+    city_choise = int(input('Выберите ваш город: '))
+    user.city = list_of_cityes['items'][city_choise - 1]
+    print('Вы выбрали ', list_of_cityes['items'][city_choise - 1])
 
 
 def get_match_users(target, api):
@@ -282,8 +145,11 @@ def save_result(data):
     print('Запись результатов в файл завершена.')
 
 
-def show_results():
-    result = get_10_users_from_db()
+def show_results(flag_db):
+    if flag_db:
+        result = get_10_users_from_db()
+    else:
+        result = get_10_users_from_file()
     while len(result) == 10:
         for user in result:
             print(user)
@@ -295,18 +161,25 @@ def show_results():
                 print('Заверешение работы программы. Удачной встречи! Будте счастливы! Всего Вам доброго!')
                 exit()
             elif user_asnwer in ['s', 'S', 'Save', 'SAVE', 'save', '[S]ave']:
+                print('Подготовка файла к записи')
+                print(type(result))
+                print(result)
                 save_result(result)
                 print('Данные сохранены! Удачной встречи! Будте счастливы! Всего Вам доброго!')
                 exit()
             elif user_asnwer == '':
-                show_results()
+                show_results(flag_db)
             else:
                 print('Неправильный ввод! Повторите!')
-    print('до новых встреч')
+    print('Пользователи в базе закончились, сохраняем оставшихся')
+    print(result)
+    save_result(result)
+    print('Данные сохранены! Удачной встречи! Будте счастливы! Всего Вам доброго!')
     exit()
 
 
 def get_10_users_from_db():
+    """ получить 10 пользователей из БД """
     users = []
     for i in range(10):
         user_ = sql.get_one()
@@ -317,6 +190,23 @@ def get_10_users_from_db():
     return users
 
 
+def get_10_users_from_file():
+    """ получить 10 пользователей из файла """
+    users = []
+    with open('fdatabase.txt', 'r', encoding='utf-8') as f:
+        text = f.readlines()
+        tmp_count = 0
+        for line in text:
+            if tmp_count == 10:
+                break
+            users.append(eval(line))
+            tmp_count += 1
+        del text[0:10]
+        with open('fdatabase.txt', 'w', encoding='utf-8') as f:
+            f.writelines(text)
+        return users
+
+
 def main():
     TOKEN = get_token()
     API = auth(TOKEN)
@@ -325,18 +215,28 @@ def main():
     target_id = get_target(API)
     target = User(target_id, API)
     get_age(target)  # получаем возраст цели
+    if not target.city:  # уточняем город цели
+        get_city(target)
     base_users = get_match_users(target, API)  # получаем список найденных пользователей
     get_top3_photos(base_users, API)  # сопоставляем три фотограифи
     sorted_users = reversed(sorted(base_users, key=operator.attrgetter('kpi')))  # сортируем по весам
-    sql.delete_tables()  # удаляем старые таблицы
-    sql.create_db(target.id)  # создаем новые таблицы
+    # - - - П Р А В К И - - -
+    if sql.get_db_status() == False:
+        db = False  # проверка доступности БД
+    else:
+        db = True
+        sql.delete_tables()  # удаляем старые таблицы
+        sql.create_db(target.id)  # создаем новые таблицы
     for i in sorted_users:
         temp = i.show_result()
-        print(temp)
+        # print(temp)
         temp_json = json.dumps(temp)
-        sql.add_user(i.id, i.kpi, temp_json)  # отправляем все, что нашли в БД
+        if db:
+            sql.add_user(i.id, i.kpi, temp_json)  # отправляем все, что нашли в БД
+        else:
+            file_base.add_user(i.id, i.kpi, temp_json)  # отправляем все, что нашли в файл
 
-    show_results()  # выводим результат
+    show_results(db)  # выводим результат
 
 
 if __name__ == '__main__':
